CityMap переделан в функциональный компонент, стейт перенесен в HOC
